Set a wrapper & pass all tests :+1:
