+++ AFAAS/core/adapters/openai/chatmodel.py
@@ -13,19 +13,19 @@ from AFAAS.core.adapters.openai.common import (
     OpenAIModelName,
     OpenAIPromptConfiguration,
     OpenAISettings,
-    _OpenAIRetryHandler,
 )
 aclient = AsyncOpenAI(api_key=os.environ["OPENAI_API_KEY"])
 
 from AFAAS.configs.schema import Configurable
 from AFAAS.interfaces.adapters.chatmodel import (
+    _RetryHandler,
     AbstractChatModelProvider,
     AbstractChatModelResponse,
     AssistantChatMessageDict,
     ChatMessage,
     CompletionModelFunction,
 )
-from AFAAS.interfaces.adapters.language_model import Embedding, ModelTokenizer
+from AFAAS.interfaces.adapters.language_model import  ModelTokenizer
 from AFAAS.lib.sdk.logger import AFAASLogger
 
 LOG = AFAASLogger(name=__name__)
@@ -46,14 +46,6 @@ class AFAASChatOpenAI(Configurable[OpenAISettings], AbstractChatModelProvider):
         self._credentials = settings.credentials
         self._budget = settings.budget
         self._chat = settings.chat
-
-        retry_handler = _OpenAIRetryHandler(
-            num_retries=self._settings.configuration.retries_per_request,
-        )
-
-        self._create_chat_completion = retry_handler(_create_chat_completion)
-
-        self._func_call_fails_count = 0
 
     def get_token_limit(self, model_name: str) -> int:
         return OPEN_AI_MODELS[model_name].max_tokens
@@ -113,94 +105,6 @@ class AFAASChatOpenAI(Configurable[OpenAISettings], AbstractChatModelProvider):
         num_tokens += 3  # every reply is primed with <|start|>assistant<|message|>
         return num_tokens
 
-    async def create_chat_completion(
-        self,
-        chat_messages: list[ChatMessage],
-        tools: list[CompletionModelFunction],
-        llm_model_name: OpenAIModelName,
-        tool_choice: str,
-        default_tool_choice: str,  # This one would be called after 3 failed attemps(cf : try/catch block)
-        completion_parser: Callable[[AssistantChatMessageDict], _T] = lambda _: None,
-        **kwargs,
-    ) -> AbstractChatModelResponse[_T]:
-        # ##############################################################################
-        # ### Step 1: Prepare arguments for API call
-        # ##############################################################################
-        completion_kwargs = self._initialize_completion_args(
-            model_name=llm_model_name,
-            tools=tools,
-            tool_choice=tool_choice,
-            **kwargs,
-        )
-
-        # ##############################################################################
-        # ### Step 2: Execute main chat completion and extract details
-        # ##############################################################################
-        response = await self._get_chat_response(
-            model_prompt=chat_messages, **completion_kwargs
-        )
-        response_message, response_args = self._extract_response_details(
-            response=response, model_name=llm_model_name
-        )
-
-        # ##############################################################################
-        # ### Step 3: Handle missing function call and retry if necessary
-        # ##############################################################################
-        if self._should_retry_function_call(
-            tools=tools, response_message=response_message
-        ):
-            if (
-                self._func_call_fails_count
-                <= self._settings.configuration.maximum_retry
-            ):
-                return await self._retry_chat_completion(
-                    model_prompt=chat_messages,
-                    tools=tools,
-                    completion_kwargs=completion_kwargs,
-                    model_name=llm_model_name,
-                    completion_parser=completion_parser,
-                    default_tool_choice=default_tool_choice,
-                    response=response,
-                    response_args=response_args,
-                )
-
-            # FIXME, TODO, NOTE: Organize application save feedback loop to improve the prompts, as it is not normal that function are not called
-            response_message["tool_calls"] = None
-            response.choices[0].message["tool_calls"] = None
-            # self._handle_failed_retry(response_message)
-
-        # ##############################################################################
-        # ### Step 4: Reset failure count and integrate improvements
-        # ##############################################################################
-        self._func_call_fails_count = 0
-
-        # ##############################################################################
-        # ### Step 5: Self feedback
-        # ##############################################################################
-
-        # Create an option to deactivate feedbacks
-        # Option : Maximum number of feedbacks allowed
-
-        # Prerequisite : Read OpenAI API (Chat Model) tool_choice section
-
-        # User : 1 shirt take 5 minutes to dry , how long take 10 shirt to dry
-        # Assistant : It takes 50 minutes
-
-        # System : "The user question was ....
-        # The Assistant Response was ..."
-        # Is it ok ?
-        # If not provide a feedback
-
-        # => T shirt can be dried at the same time
-
-        # ##############################################################################
-        # ### Step 6: Formulate the response
-        # ##############################################################################
-        return self._formulate_final_response(
-            response_message=response_message,
-            completion_parser=completion_parser,
-            response_args=response_args,
-        )
 
     def _initialize_completion_args(
         self,
@@ -213,12 +117,6 @@ class AFAASChatOpenAI(Configurable[OpenAISettings], AbstractChatModelProvider):
         completion_kwargs["tool_choice"] = tool_choice
         return completion_kwargs
 
-    async def _get_chat_response(
-        self, model_prompt: list[ChatMessage], **completion_kwargs: Any
-    ) -> AsyncCompletions:
-        return await self._create_chat_completion(
-            messages=model_prompt, **completion_kwargs
-        )
 
     def _extract_response_details(
         self, response: AsyncCompletions, model_name: str
@@ -241,48 +139,6 @@ class AFAASChatOpenAI(Configurable[OpenAISettings], AbstractChatModelProvider):
             return True
         return False
 
-    async def _retry_chat_completion(
-        self,
-        model_prompt: list[ChatMessage],
-        tools: list[CompletionModelFunction],
-        completion_kwargs: Dict[str, Any],
-        model_name: str,
-        completion_parser: Callable[[AssistantChatMessageDict], _T],
-        default_tool_choice: str,
-        response: AsyncCompletions,
-        response_args: Dict[str, Any],
-    ) -> AbstractChatModelResponse[_T]:
-        completion_kwargs = self._update_function_call_for_retry(
-            completion_kwargs=completion_kwargs,
-            default_tool_choice=default_tool_choice,
-        )
-        completion_kwargs["tools"] = tools
-        response.update(response_args)
-        self._budget.update_usage_and_cost(model_response=response)
-        return await self.create_chat_completion(
-            chat_messages=model_prompt,
-            llm_model_name=model_name,
-            completion_parser=completion_parser,
-            **completion_kwargs,
-        )
-
-    def _update_function_call_for_retry(
-        self, completion_kwargs: Dict[str, Any], default_tool_choice: str
-    ) -> Dict[str, Any]:
-        if (
-            self._func_call_fails_count
-            >= self._settings.configuration.maximum_retry_before_default_function
-        ):
-            completion_kwargs["tool_calls"] = default_tool_choice
-        else:
-            completion_kwargs["tool_calls"] = completion_kwargs.get(
-                "tool_calls", "auto"
-            )
-        completion_kwargs["default_tool_choice"] = completion_kwargs.get(
-            "default_tool_choice", default_tool_choice
-        )
-        self._func_call_fails_count += 1
-        return completion_kwargs
 
     # def _handle_failed_retry(self, response_message: Dict[str, Any], response: openai.Completion) -> None:
 
@@ -299,12 +155,6 @@ class AFAASChatOpenAI(Configurable[OpenAISettings], AbstractChatModelProvider):
         )
         self._budget.update_usage_and_cost(model_response=response)
         return response
-
-    async def create_language_completion(self, **kwargs):
-        LOG.warning(
-            "create_language_completion is deprecated, use create_chat_completion"
-        )
-        return await self.create_chat_completion(**kwargs)
 
     def _get_completion_kwargs(
         self,
@@ -334,40 +184,32 @@ class AFAASChatOpenAI(Configurable[OpenAISettings], AbstractChatModelProvider):
     def get_default_config(self) -> OpenAIPromptConfiguration:
         return OPEN_AI_DEFAULT_CHAT_CONFIGS.SMART_MODEL_32K
 
-
-async def _create_chat_completion(
-    messages: list[ChatMessage], *_, **kwargs
-) -> AsyncCompletions:
-
-    raw_messages = [
-        message.dict(include={"role", "content", "tool_calls", "name"})
-        for message in messages
-    ]
-
-    if not "tools" in kwargs or kwargs["tools"] is None or len(kwargs["tools"]) == 0:
-        if "tools" in kwargs:
-            del kwargs["tools"]
-        kwargs.pop("tool_choice", None)
-
-    else:
-        # kwargs["tools"] = [function for function in kwargs["tools"]]
-        if len(kwargs["tools"]) == 0:
-            del kwargs["tools"]
+    def make_chat_kwargs(self, **kwargs) -> dict:          
+        if not "tools" in kwargs or kwargs["tools"] is None or len(kwargs["tools"]) == 0:
+            if "tools" in kwargs:
+                del kwargs["tools"]
             kwargs.pop("tool_choice", None)
-        elif len(kwargs["tools"]) == 1:
-            kwargs["tool_choice"] = {
-                "type": "function",
-                "function": {"name": kwargs["tools"][0]["function"]["name"]},
-            }
-        elif kwargs["tool_choice"] != "auto":
-            kwargs["tool_choice"] = {
-                "type": "function",
-                "function": {"name": kwargs["tool_choice"]},
-            }
 
-    LOG.trace(raw_messages[0]["content"])
-    LOG.trace(kwargs)
-    return_value = await aclient.chat.completions.create(
-        messages=raw_messages, **kwargs
-    )
-    return return_value
+        else:
+            if len(kwargs["tools"]) == 0:
+                del kwargs["tools"]
+                kwargs.pop("tool_choice", None)
+            elif len(kwargs["tools"]) == 1:
+                kwargs["tool_choice"] = self.make_tool_choice(name= kwargs["tools"][0]["function"]["name"])
+            elif kwargs["tool_choice"] != "auto":
+                kwargs["tool_choice"] = self.make_tool_choice(name= kwargs["tool_choice"])
+        return kwargs
+
+    def make_tool_choice(self , name : str) -> dict:
+        return {
+            "type": "function",
+            "function": {"name": name},
+        }
+
+    async def chat(
+        self, messages: list[ChatMessage], *_, **llm_kwargs
+    ) -> AsyncCompletions:
+        self.llm_model = aclient.chat
+        return await aclient.chat.completions.create(
+            messages=messages, **llm_kwargs
+        )
