+++ AFAAS/core/adapters/openai/common.py
@@ -1,11 +1,9 @@
 import enum
-import functools
 import math
 import os
-import time
 from typing import Callable, ClassVar, ParamSpec, TypeVar
 
-from openai import APIError, AsyncOpenAI, RateLimitError
+from openai import AsyncOpenAI
 
 from AFAAS.configs.schema import Field
 from AFAAS.interfaces.adapters.chatmodel import (
@@ -175,66 +173,6 @@ class OpenAISettings(BaseModelProviderSettings):
 
     name : str =  "chat_model_provider"
     description : str =  "Provides access to OpenAI's API."
-
-
-class _OpenAIRetryHandler:
-    """Retry Handler for OpenAI API call.
-
-    Args:
-        num_retries int: Number of retries. Defaults to 10.
-        backoff_base float: Base for exponential backoff. Defaults to 2.
-        warn_user bool: Whether to warn the user. Defaults to True.
-    """
-
-    _retry_limit_msg = "Error: Reached rate limit, passing..."
-    _api_key_error_msg = (
-        "Please double check that you have setup a PAID OpenAI API Account. You can "
-        "read more here: https://docs.agpt.co/setup/#getting-an-api-key"
-    )
-    _backoff_msg = "Error: API Bad gateway. Waiting {backoff} seconds..."
-
-    def __init__(
-        self,
-        num_retries: int = 10,
-        backoff_base: float = 2.0,
-        warn_user: bool = True,
-    ):
-        self._num_retries = num_retries
-        self._backoff_base = backoff_base
-        self._warn_user = warn_user
-
-    def _log_rate_limit_error(self) -> None:
-        LOG.trace(self._retry_limit_msg)
-        if self._warn_user:
-            LOG.warning(self._api_key_error_msg)
-            self._warn_user = False
-
-    def _backoff(self, attempt: int) -> None:
-        backoff = self._backoff_base ** (attempt + 2)
-        LOG.trace(self._backoff_msg.format(backoff=backoff))
-        time.sleep(backoff)
-
-    def __call__(self, func: Callable[_P, _T]) -> Callable[_P, _T]:
-        @functools.wraps(func)
-        async def _wrapped(*args: _P.args, **kwargs: _P.kwargs) -> _T:
-            num_attempts = self._num_retries + 1  # +1 for the first attempt
-            for attempt in range(1, num_attempts + 1):
-                try:
-                    return await func(*args, **kwargs)
-
-                except RateLimitError:
-                    if attempt == num_attempts:
-                        raise
-                    self._log_rate_limit_error()
-
-                except APIError as e:
-                    if (e.http_status != 502) or (attempt == num_attempts):
-                        raise
-                except Exception as e:
-                    LOG.warning(e)
-                self._backoff(attempt)
-
-        return _wrapped
 
 
 class OpenAIPromptConfiguration(AbstractPromptConfiguration):
